4-4-2020

+++ fDatPhong.cs
@@ -29,6 +29,101 @@ namespace KS
             datPhongDAO = new DatPhongDAO();
         }
 
+        bool checkDatPhongHopLe()
+        {
+            if (txbTenkh.Text == "") lbtenkh.Visible = true;
+            else lbtenkh.Visible = false;
+            if (txbGioiTinh.Text == "") lbgioitinh.Visible = true;
+            else lbgioitinh.Visible = false;
+            if (txbDiaChi.Text == "") lbdiachi.Visible = true;
+            else lbdiachi.Visible = false;
+            if (txbSDT.Text == "") lbSDT.Visible = true;
+            else lbSDT.Visible = false;
+            if (txbCMT.Text == "") lbCMT.Visible = true;
+            else lbCMT.Visible = false;
+            if (cbbChonLoaiPhg.Text == "") lbChonLoaiPhong.Visible = true;
+            else lbChonLoaiPhong.Visible = false;
+            if (cbbChonPhg.Text == "") lbChonPhong.Visible = true;
+            else lbChonPhong.Visible = false;
+            if (txbTenkh.Text != "" && txbGioiTinh.Text != "" && txbDiaChi.Text != "" && txbSDT.Text != "" && txbCMT.Text != "") return true;
+            return false;
+        }
+
+        bool checkThoiGianDatPhongHopLe(DatPhong datPhong, int kind)
+        {
+            DatPhongDAO datPhongDAO = new DatPhongDAO();
+            var datPhg = datPhongDAO.GetDatPhong();
+            if (dtPKerNgayVao.Value.Date > dtPKerNgayRa.Value.Date)
+            {
+                MessageBox.Show("Không được ngày ở > ngày đi");
+                return false;
+            }
+            
+            foreach (var phg in datPhg)
+            {
+                if (kind == 1)
+                {
+                    if (phg.MAPHONG == Convert.ToInt32(cbbChonPhg.Text))
+                    {
+                        if ((dtPKerNgayVao.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayVao.Value.Date <= phg.NGAYDI.Value.Date) || (dtPKerNgayRa.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayRa.Value.Date <= phg.NGAYDI.Value.Date))
+                        {
+                            MessageBox.Show("Không được trùng thời gian với đơn đặt cùng phòng");
+                            return false;
+                        }
+                        if (dtPKerNgayVao.Value.Date < DateTime.Today || dtPKerNgayRa.Value.Date < DateTime.Today)
+                        {
+                            MessageBox.Show("Không được đặt thời gian trong quá khứ");
+                            return false;
+                        }
+                    }
+                }
+                if (kind == 2)
+                {
+                    if (phg.MAPHONG == Convert.ToInt32(cbbChonPhg.Text))
+                    {
+                        if (phg.MADATPHONG == datPhong.MADATPHONG && ((dtPKerNgayVao.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayVao.Value.Date <= phg.NGAYDI.Value.Date) || (dtPKerNgayRa.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayRa.Value.Date <= phg.NGAYDI.Value.Date))) continue;
+                        if ((dtPKerNgayVao.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayVao.Value.Date <= phg.NGAYDI.Value.Date) || (dtPKerNgayRa.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayRa.Value.Date <= phg.NGAYDI.Value.Date))
+                            return false;
+                    }
+                }
+            }
+            return true;
+        }
+
+        void loadLoaiPhong()
+        {
+            LoaiPhongDAO loaiPhongDAO = new LoaiPhongDAO();
+            var lstLoaiPhong = loaiPhongDAO.GetLoaiPhong();
+            cbbChonLoaiPhg.DataSource = lstLoaiPhong;
+            cbbChonLoaiPhg.DisplayMember = "TenLoaiPhong";
+        }
+
+        void RefreshForm()
+        {
+            loadLoaiPhong();
+            loadThongTinPhongDangChon();
+        }
+
+        void editDatPhong()
+        {
+            int id = int.Parse(dtgvThongTinPhg.SelectedCells[0].OwningRow.Cells["maDatPhong"].Value.ToString());
+            DatPhong datPhong = db.DatPhongs.Find(id);
+            datPhong.MAPHONG = int.Parse(cbbChonPhg.Text);
+            datPhong.NGAYO = dtPKerNgayVao.Value.Date;
+            datPhong.NGAYDI = dtPKerNgayRa.Value.Date;
+            datPhong.GiaPhongHienTai = Convert.ToDouble(txbgiaPhong.Text);
+            datPhong.TRATRUOC = Convert.ToDouble(txbTraTruoc.Text);
+            if (checkThoiGianDatPhongHopLe(datPhong, 2) == true)
+            {
+                if (MessageBox.Show("Bạn có thật sự muốn Sửa đặt phòng này?", "Thông Báo", MessageBoxButtons.OKCancel) == System.Windows.Forms.DialogResult.OK)
+                {
+                    db.SaveChanges();
+                    loadThongTinPhongDangChon();
+                    MessageBox.Show("Sửa thành công");
+                }
+            }
+
+        }
         private void CreateButtonDelete()
         {
             DataGridViewButtonColumn buttonXoa = new DataGridViewButtonColumn();
@@ -45,8 +140,8 @@ namespace KS
         {
             var result = from c in db.Phongs
                          from a in db.LoaiPhongs
-
                          where c.MALOAIPHONG == a.MALOAIPHONG
+                         && c.isDelete == false
                          select new
                          {
                              c.MAPHONG,
@@ -57,20 +152,20 @@ namespace KS
                              c.DONVITIENTE
                          };
             dtgvDSPhg.DataSource = result.ToList();
-        }
-
+        }//ok
         void loadThongTinPhongDat()
         {
             var result = (from a in db.Phongs
                           from b in db.LoaiPhongs
                           where a.MALOAIPHONG == b.MALOAIPHONG
                           && b.TENLOAIPHONG == cbbChonLoaiPhg.Text
+                          && a.isDelete == false
                           select new
                           {
                               tenPhong = a.MAPHONG.ToString(),
                               giaPhong = a.GIAPHONG
                           }).ToList();
-            var lstPhong = phongDAO.GetPhong();
+            var lstPhong = phongDAO.GetPhongNotDeleted();
             if (result.Count > 0)
             {
                 foreach (var value in result)
@@ -79,8 +174,6 @@ namespace KS
                 }
                 cbbChonPhg.Text = result[0].tenPhong;
             }
-            dtPKerNgayVao.MinDate = DateTime.Today;
-            dtPKerNgayRa.MinDate = DateTime.Today;
         }
         void loadThongTinPhongDangChon()
         {
@@ -88,7 +181,7 @@ namespace KS
                           from b in db.DatPhongs
                           where a.MAPHONG == b.MAPHONG &&
                           cbbChonPhg.Text == a.MAPHONG.ToString()
-                          && b.isDelete ==false
+                          && b.isDelete == false
                           select new RowDatPhong
                           {
                               MaDatPhong = b.MADATPHONG,
@@ -105,85 +198,61 @@ namespace KS
 
         private void btnDatPhg_Click(object sender, EventArgs e)
         {
-            if (txbTenkh.Text == "") lbtenkh.Visible = true;
-            else lbtenkh.Visible = false;
-            if (txbGioiTinh.Text == "") lbgioitinh.Visible = true;
-            else lbgioitinh.Visible = false;
-            if (txbDiaChi.Text == "") lbdiachi.Visible = true;
-            else lbdiachi.Visible = false;
-            if (txbSDT.Text == "") lbSDT.Visible = true;
-            else lbSDT.Visible = false;
-            if (txbCMT.Text == "") lbCMT.Visible = true;
-            else lbCMT.Visible = false;
-            if (cbbChonLoaiPhg.Text == "") lbChonLoaiPhong.Visible = true;
-            else lbChonLoaiPhong.Visible = false;
-            if (cbbChonPhg.Text == "") lbChonPhong.Visible = true;
-            else lbChonPhong.Visible = false;
             var datPhg = datPhongDAO.GetDatPhong();
-            if (cbbChonPhg.Text != "" && cbbChonLoaiPhg.Text != "" && txbTenkh.Text != "" && txbGioiTinh.Text != "" && txbDiaChi.Text != "" && txbSDT.Text != "" && txbCMT.Text != "")
+            KhachHang khachHang = new KhachHang()
             {
-                foreach (var phg in datPhg)
+                TENKH = txbTenkh.Text,
+                GIOITINH = txbGioiTinh.Text,
+                NGAYSINH = datimeNgaySinh.Value.Date,
+                DIACHI = txbDiaChi.Text,
+                SODIENTHOAI = txbSDT.Text,
+                CHUNGMINHTHU = txbCMT.Text
+            };
+            if (checkDatPhongHopLe() == true)
+            {
+                var kh = khachHangDAO.GetKhachHang(txbCMT.Text);
+                DatPhong datPhong = new DatPhong()
                 {
-                    if (phg.MAPHONG == Convert.ToInt32(cbbChonPhg.Text)){
-                        if((dtPKerNgayVao.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayVao.Value.Date <= phg.NGAYDI.Value.Date)|| (dtPKerNgayRa.Value.Date >= phg.NGAYO.Value.Date && dtPKerNgayRa.Value.Date <= phg.NGAYDI.Value.Date))
+                    MAPHONG = Convert.ToInt32(cbbChonPhg.Text),
+                    MAKH = kh.MAKH,
+                    TRATRUOC = Convert.ToDouble(txbTraTruoc.Text),
+                    NGAYO = dtPKerNgayVao.Value.Date,
+                    NGAYDI = dtPKerNgayRa.Value.Date,
+                    TrangThaiThanhToan = "Trả trước 20%",
+                    GiaPhongHienTai = Convert.ToDouble(txbgiaPhong.Text),
+                    isDelete = false
+                };
+
+                if (checkThoiGianDatPhongHopLe(datPhong, 1) == true)
+                {
+                    if (MessageBox.Show("Bạn có thật sự muốn đặt phòng này?", "Thông Báo", MessageBoxButtons.OKCancel) == System.Windows.Forms.DialogResult.OK)
+                    {
+                        KhachHangDAO khachHangDAO = new KhachHangDAO();
+                        if (khachHangDAO.isKhachHangHopLe(khachHang) == false)//ko hop le
                         {
-                            MessageBox.Show("Đặt ko thành công!Trùng thời gian ", "Thông Báo");
+                            MessageBox.Show("CMT hoặc SĐT tồn tại");
                             return;
                         }
-                    }
-                }
-            }
-            if (txbTenkh.Text != "" && txbGioiTinh.Text != "" && txbDiaChi.Text != "" && txbSDT.Text != "" && txbCMT.Text != "")
-            {
-                if(dtPKerNgayVao.Value.Date > dtPKerNgayRa.Value.Date)
-                {
-                    MessageBox.Show("Không thành công! Ngày ở sau ngày đi");
-                }
-                else if (MessageBox.Show("Bạn có thật sự muốn đặt phòng này?", "Thông Báo", MessageBoxButtons.OKCancel) == System.Windows.Forms.DialogResult.OK)
-                {
-                    KhachHang khachHang = new KhachHang()
-                    {
-                        TENKH = txbTenkh.Text,
-                        GIOITINH = txbGioiTinh.Text,
-                        NGAYSINH = datimeNgaySinh.Value.Date,
-                        DIACHI = txbDiaChi.Text,
-                        SODIENTHOAI = txbSDT.Text,
-                        CHUNGMINHTHU = txbCMT.Text
-                    };
-                    if (db.KhachHangs.SingleOrDefault(x => x.CHUNGMINHTHU == khachHang.CHUNGMINHTHU) == null)
-                    {
-                        db.KhachHangs.Add(khachHang);
+                        if (khachHangDAO.isKhachHangTonTai(khachHang) == false)//ko ton tai
+                        {
+                            db.KhachHangs.Add(khachHang);
+                            db.SaveChanges();
+                        }
+                        db.DatPhongs.Add(datPhong);
+                        MessageBox.Show("Đặt phòng thành công", "Thông báo");
                         db.SaveChanges();
+                        loadThongTinPhongDangChon();
                     }
-                    var kh = khachHangDAO.GetKhachHang(txbCMT.Text);
-                    DatPhong datPhong = new DatPhong()
-                    {
-                        MAPHONG = Convert.ToInt32(cbbChonPhg.Text),
-                        MAKH = kh.MAKH,
-                        TRATRUOC = Convert.ToDouble(txbTraTruoc.Text),
-                        NGAYO = dtPKerNgayVao.Value.Date,
-                        NGAYDI = dtPKerNgayRa.Value.Date,
-                        TrangThaiThanhToan = "Trả trước 20%",
-                        GiaPhongHienTai = Convert.ToDouble(txbgiaPhong.Text),
-                        isDelete = false
-                    };
-                    db.DatPhongs.Add(datPhong);
-                    MessageBox.Show("Đặt phòng thành công", "Thông báo");
-                    db.SaveChanges();
-                    loadThongTinPhongDangChon();
                 }
             }
         }
 
         private void fAddNewKH_Load(object sender, EventArgs e)
         {
-            cbbChonLoaiPhg.Text = "Tiêu Chuẩn";
-            cbbChonPhg.Items.Clear();
             loadThongTinPhongDat();
-            loadThongTinPhongDangChon();
             loadDanhSachPhong();
+            RefreshForm();
             CreateButtonDelete();
-            
         }
 
         private void txbSDT_KeyUp(object sender, KeyEventArgs e)
@@ -226,7 +295,7 @@ namespace KS
         {
             int maphg = int.Parse(cbbChonPhg.Text);
             var phong = phongDAO.GetPhong(maphg);
-            
+
             var tongNgayo = (dtPKerNgayRa.Value.Date - dtPKerNgayVao.Value.Date).TotalDays;
             if (tongNgayo < 0) MessageBox.Show("Chọn ngày ở <ngày đi");
             else
@@ -255,24 +324,21 @@ namespace KS
                 e.Handled = true;
         }
 
-        private void dtgvThongTinPhg_SelectionChanged(object sender, EventArgs e)
-        {
-        }
-
         private void dtgvThongTinPhg_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
             var senderGrid = (DataGridView)sender;
-
             if (senderGrid.Columns[e.ColumnIndex] is DataGridViewButtonColumn &&
                 e.RowIndex >= 0)
             {
-                //ham tim theo id hoac object, ham sua , ham xoa reomve theo list object hoac remomve theo object
-                var selectedDatPhong = db.DatPhongs.Find(((RowDatPhong)dtgvThongTinPhg.CurrentRow.DataBoundItem).MaDatPhong);
-                db.DatDichVus.RemoveRange(selectedDatPhong.DatDichVus.ToList());
-                db.DatPhongs.Remove(selectedDatPhong);
-                db.SaveChanges();
-                MessageBox.Show("remove success");
-                loadThongTinPhongDangChon();
+                if (MessageBox.Show("Bạn có thật sự muốn hủy đặt phòng này?", "Thông Báo", MessageBoxButtons.OKCancel) == System.Windows.Forms.DialogResult.OK)
+                {
+                    var selectedDatPhong = db.DatPhongs.Find(((RowDatPhong)dtgvThongTinPhg.CurrentRow.DataBoundItem).MaDatPhong);
+                    db.DatDichVus.RemoveRange(selectedDatPhong.DatDichVus.ToList());
+                    db.DatPhongs.Remove(selectedDatPhong);
+                    db.SaveChanges();
+                    MessageBox.Show("remove success");
+                    loadThongTinPhongDangChon();
+                }
             }
         }
 
@@ -281,14 +347,14 @@ namespace KS
             loadTraTruoc();
         }
 
-        private void fDatPhong_FormClosing(object sender, FormClosingEventArgs e)
-        {
-            f.loadSoDoPhong();
-        }
-
         private void dtPKerNgayVao_ValueChanged(object sender, EventArgs e)
         {
             loadTraTruoc();
+        }
+
+        private void btnSuaDatPhong_Click(object sender, EventArgs e)
+        {
+            editDatPhong();
         }
     }
 }
